@@ -1,5 +1,7 @@
 apply plugin: 'com.android.library'
 
+apply plugin: 'me.tatarka.retrolambda'
+
 android {
     compileSdkVersion 21
     buildToolsVersion '21.1.2'
@@ -10,8 +12,8 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
     buildTypes {
         release {
@@ -23,7 +25,8 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:support-v4:21.0.2'
+    compile 'com.android.support:support-v4:21.0.3'
     compile 'com.squareup.okhttp:okhttp:2.1.0'
+    compile 'com.afollestad:material-dialogs:0.6.3.5'
     compile project(':Afw')
 }
