@@ -39,9 +39,12 @@ import android.widget.FrameLayout;
 import android.widget.ImageButton;
 import android.widget.LinearLayout;
 import android.widget.ListAdapter;
+import android.widget.ListView;
 import android.widget.Switch;
 import android.widget.TextView;
 import android.widget.Toast;
+import com.afollestad.materialdialogs.AlertDialogWrapper;
+import com.afollestad.materialdialogs.MaterialDialog;
 import org.json.JSONException;
 import org.json.JSONObject;
 import yuku.afw.V;
@@ -591,10 +594,10 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 				res.selectVerseCount = selectVerseCount;
 				return res;
 			} else {
-				new AlertDialog.Builder(this)
-				.setMessage("Invalid ari: " + ari)
-				.setPositiveButton(R.string.ok, null)
-				.show();
+				new AlertDialogWrapper.Builder(this)
+					.setMessage("Invalid ari: " + ari)
+					.setPositiveButton(R.string.ok, null)
+					.show();
 				return null;
 			}
 		} else if (intent.hasExtra("lid")) {
@@ -608,10 +611,10 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 				res.selectVerseCount = selectVerseCount;
 				return res;
 			} else {
-				new AlertDialog.Builder(this)
-				.setMessage("Invalid lid: " + lid)
-				.setPositiveButton(R.string.ok, null)
-				.show();
+				new AlertDialogWrapper.Builder(this)
+					.setMessage("Invalid lid: " + lid)
+					.setPositiveButton(R.string.ok, null)
+					.show();
 				return null;
 			}
 		} else {
@@ -742,7 +745,7 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 		} catch (Throwable e) { // so we don't crash on the beginning of the app
 			Log.e(TAG, "Error opening main version", e);
 
-			new AlertDialog.Builder(IsiActivity.this)
+			new AlertDialogWrapper.Builder(IsiActivity.this)
 				.setMessage(getString(R.string.version_error_opening, mv.longName))
 				.setPositiveButton(R.string.ok, null)
 				.show();
@@ -767,7 +770,7 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 		} catch (Throwable e) { // so we don't crash on the beginning of the app
 			Log.e(TAG, "Error opening split version", e);
 
-			new AlertDialog.Builder(IsiActivity.this)
+			new AlertDialogWrapper.Builder(IsiActivity.this)
 				.setMessage(getString(R.string.version_error_opening, mv.longName))
 				.setPositiveButton(R.string.ok, null)
 				.show();
@@ -1002,15 +1005,24 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 	
 	void bGoto_longClick() {
 		if (history.getSize() > 0) {
-			new AlertDialog.Builder(this)
-			.setAdapter(historyAdapter, (dialog, which) -> {
-				int ari = history.getAri(which);
+			final MaterialDialog dialog = new MaterialDialog.Builder(this)
+				.adapter(historyAdapter)
+				.autoDismiss(true)
+				.negativeText(R.string.cancel)
+				.show();
+
+			final ListView listView = dialog.getListView();
+			if (listView == null) {
+				throw new RuntimeException("ListView should not be null");
+			}
+
+			listView.setOnItemClickListener((parent, view, position, id) -> {
+				dialog.dismiss();
+				int ari = history.getAri(position);
 				jumpToAri(ari, true);
 				history.add(ari);
 				Preferences.setBoolean(Prefkey.history_button_understood, true);
-			})
-			.setNegativeButton(R.string.cancel, null)
-			.show();
+			});
 		} else {
 			Toast.makeText(this, R.string.recentverses_not_available, Toast.LENGTH_SHORT).show();
 		}
@@ -1635,12 +1647,15 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 			if (markers.size() == 1) {
 				openBookmarkDialog(markers.get(0)._id);
 			} else {
-                AlertDialog dialog = new AlertDialog.Builder(IsiActivity.this)
+                AlertDialog dialog = new AlertDialogWrapper.Builder(IsiActivity.this)
                     .setTitle(R.string.edit_bookmark)
-                    .setAdapter(new MultipleMarkerSelectAdapter(markers, Marker.Kind.bookmark), (_dialog_, which) -> openBookmarkDialog(markers.get(which)._id))
+                    .setAdapter(new MultipleMarkerSelectAdapter(markers, Marker.Kind.bookmark))
                     .show();
-                dialog.getListView().setDrawSelectorOnTop(true);
-            }
+
+				final ListView listView = dialog.getListView();
+				listView.setDrawSelectorOnTop(true);
+				listView.setOnItemClickListener((parent, view, position, id) -> openBookmarkDialog(markers.get(position)._id));
+			}
         }
 
 		void openNoteDialog(final long _id) {
@@ -1796,21 +1811,21 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 							VerseRenderer.appendSuperscriptNumber(footnoteText, arif & 0xff);
 							footnoteText.append(" ");
 
-							new AlertDialog.Builder(IsiActivity.this)
-							.setMessage(FormattedTextRenderer.render(fe.content, footnoteText))
-							.setPositiveButton("OK", null)
-							.show();
+							new AlertDialogWrapper.Builder(IsiActivity.this)
+								.setMessage(FormattedTextRenderer.render(fe.content, footnoteText))
+								.setPositiveButton(R.string.ok, null)
+								.show();
 						} else {
-							new AlertDialog.Builder(IsiActivity.this)
-							.setMessage(String.format(Locale.US, "Error: footnote arif 0x%08x couldn't be loaded", arif))
-							.setPositiveButton("OK", null)
-							.show();
+							new AlertDialogWrapper.Builder(IsiActivity.this)
+								.setMessage(String.format(Locale.US, "Error: footnote arif 0x%08x couldn't be loaded", arif))
+								.setPositiveButton(R.string.ok, null)
+								.show();
 						}
 					} else {
-						new AlertDialog.Builder(IsiActivity.this)
-						.setMessage("Error: Unknown inline link type: " + type)
-						.setPositiveButton("OK", null)
-						.show();
+						new AlertDialogWrapper.Builder(IsiActivity.this)
+							.setMessage("Error: Unknown inline link type: " + type)
+							.setPositiveButton("OK", null)
+							.show();
 					}
 				}
 			};
@@ -2197,7 +2212,7 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 			final ProgressMarkListDialog dialog = new ProgressMarkListDialog();
 			dialog.show(getSupportFragmentManager(), "dialog_progress_mark_list");
 		} else {
-			new AlertDialog.Builder(this)
+			new AlertDialogWrapper.Builder(this)
 				.setMessage(R.string.pm_activate_tutorial)
 				.setPositiveButton(R.string.ok, null)
 				.show();
@@ -2217,7 +2232,7 @@ public class IsiActivity extends BaseLeftDrawerActivity implements XrefDialog.Xr
 			jumpToAri(ari, false);
 			history.add(ari);
 		} else {
-			new AlertDialog.Builder(this)
+			new AlertDialogWrapper.Builder(this)
 				.setMessage(R.string.pm_activate_tutorial)
 				.setPositiveButton(R.string.ok, null)
 				.show();
