@@ -1,6 +1,5 @@
 package yuku.alkitab.base.ac;
 
-import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
@@ -11,9 +10,7 @@ import android.view.MenuItem;
 import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.ViewFlipper;
-
-import java.util.Date;
-
+import com.afollestad.materialdialogs.AlertDialogWrapper;
 import yuku.afw.V;
 import yuku.alkitab.base.App;
 import yuku.alkitab.base.S;
@@ -27,6 +24,8 @@ import yuku.alkitab.util.IntArrayList;
 import yuku.alkitabconverter.util.DesktopVerseFinder;
 import yuku.alkitabconverter.util.DesktopVerseParser;
 import yuku.alkitabintegration.display.Launcher;
+
+import java.util.Date;
 
 public class NoteActivity extends BaseActivity {
 	//region static constructors
@@ -122,7 +121,7 @@ public class NoteActivity extends BaseActivity {
 
 		final IntArrayList verseRanges = DesktopVerseParser.verseStringToAri(verse);
 		if (verseRanges == null || verseRanges.size() == 0) {
-			new AlertDialog.Builder(widget.getContext())
+			new AlertDialogWrapper.Builder(widget.getContext())
 				.setMessage(R.string.note_activity_cannot_parse_verse)
 				.setPositiveButton(R.string.ok, null)
 				.show();
@@ -206,7 +205,7 @@ public class NoteActivity extends BaseActivity {
 			case R.id.menuDelete: {
 				// if it's indeed not exist, check if we have some text, if we do, prompt first
 				if (marker != null || tCaption.length() > 0) {
-					new AlertDialog.Builder(this)
+					new AlertDialogWrapper.Builder(this)
 						.setMessage(R.string.anda_yakin_mau_menghapus_catatan_ini)
 						.setPositiveButton(R.string.delete, (dialog, which) -> {
 							if (marker != null) {
