@@ -28,6 +28,7 @@ import android.view.Window;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import android.widget.Toast;
+import com.afollestad.materialdialogs.AlertDialogWrapper;
 import yuku.afw.V;
 import yuku.afw.storage.Preferences;
 import yuku.alkitab.base.App;
@@ -311,7 +312,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 						final Activity activity = activityRef.get();
 						if (activity != null) {
 							if (!activity.isFinishing()) {
-								new AlertDialog.Builder(activity)
+								new AlertDialogWrapper.Builder(activity)
 									.setMessage(activity.getString(R.string.song_player_error_description, what, extra))
 									.setPositiveButton(R.string.ok, null)
 									.show();
@@ -566,7 +567,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		} return true;
 
         case R.id.menuUpdateBook: {
-			new AlertDialog.Builder(this)
+			new AlertDialogWrapper.Builder(this)
 				.setMessage(TextUtils.expandTemplate(getString(R.string.sn_update_book_explanation), currentBookName))
 				.setPositiveButton(R.string.sn_update_book_confirm_button, (dialog, which) -> updateSongBook())
 				.setNegativeButton(R.string.cancel, null)
@@ -574,7 +575,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		} return true;
 
 		case R.id.menuDeleteAll: {
-			new AlertDialog.Builder(this)
+			new AlertDialogWrapper.Builder(this)
 				.setMessage(R.string.sn_delete_all_songs_explanation)
 				.setPositiveButton(R.string.delete, (dialog, which) -> deleteAllSongs())
 				.setNegativeButton(R.string.cancel, null)
@@ -597,7 +598,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 			@Override
 			public void onFailedOrCancelled(SongBookUtil.SongBookInfo songBookInfo, Exception e) {
 				if (e != null) {
-					new AlertDialog.Builder(SongViewActivity.this)
+					new AlertDialogWrapper.Builder(SongViewActivity.this)
 						.setMessage(e.getClass().getSimpleName() + ' ' + e.getMessage())
 						.setPositiveButton(R.string.ok, null)
 						.show();
@@ -623,7 +624,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
                 runOnUiThread(() -> {
                     pd.dismiss();
 
-					new AlertDialog.Builder(SongViewActivity.this)
+					new AlertDialogWrapper.Builder(SongViewActivity.this)
 						.setMessage(getString(R.string.sn_delete_all_songs_result, count))
 						.setPositiveButton(R.string.ok, null)
 						.show()
@@ -1001,7 +1002,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 			SongBookUtil.downloadSongBook(SongViewActivity.this, songBookInfo, new SongBookUtil.OnDownloadSongBookListener() {
 				@Override public void onFailedOrCancelled(SongBookUtil.SongBookInfo songBookInfo, Exception e) {
 					if (e != null) {
-						new AlertDialog.Builder(SongViewActivity.this)
+						new AlertDialogWrapper.Builder(SongViewActivity.this)
 							.setMessage(e.getClass().getSimpleName() + ' ' + e.getMessage())
 							.setPositiveButton(R.string.ok, null)
 							.show();
